+++ TShockAPI/DB/BanManager.cs
@@ -22,6 +22,7 @@ using System.Collections.Generic;
 using System.Data;
 using MySql.Data.MySqlClient;
 using System.Collections.ObjectModel;
+using TShockAPI.DB.Queries;
 
 namespace TShockAPI.DB
 {
