+++ TShockAPI/DB/RegionManager.cs
@@ -23,6 +23,7 @@ using System.Linq;
 using MySql.Data.MySqlClient;
 using Terraria;
 using Microsoft.Xna.Framework;
+using TShockAPI.DB.Queries;
 
 namespace TShockAPI.DB
 {
