+++ TShockAPI/DB/SqlTable.cs
@@ -21,6 +21,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using MySql.Data.MySqlClient;
+using TShockAPI.DB.Queries;
 
 namespace TShockAPI.DB
 {
