+++ TShockAPI/DB/WarpsManager.cs
@@ -24,6 +24,7 @@ using System.Linq;
 using MySql.Data.MySqlClient;
 using Terraria;
 using Microsoft.Xna.Framework;
+using TShockAPI.DB.Queries;
 
 namespace TShockAPI.DB
 {
