+++ TShockAPI/Extensions/StringExt.cs
@@ -24,9 +24,9 @@ namespace TShockAPI
 	public static class StringExt
 	{
 		//Can't name it Format :(
-		public static String SFormat(this String str, params object[] args)
+		public static string SFormat(this string str, params object[] args)
 		{
-			return String.Format(str, args);
+			return string.Format(str, args);
 		}
 
 		/// <summary>
